merged: r24882,r24883,r24889 back to trunk.
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/trunk@24903 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ xbmc/cores/DummyVideoPlayer.cpp
@@ -229,7 +229,7 @@ float CDummyVideoPlayer::GetSubTitleDelay()
 void CDummyVideoPlayer::SeekTime(__int64 iTime)
 {
   m_clock = iTime;
-  m_callback.OnPlayBackSeek(iTime);
+  m_callback.OnPlayBackSeek((int)iTime);
 }
 
 // return the time in milliseconds
