+++ xbmc/cores/paplayer/PAPlayer.cpp
@@ -861,7 +861,7 @@ void PAPlayer::SeekTime(__int64 iTime /*=0*/)
   if (m_currentFile->m_lStartOffset)
     iTime += m_currentFile->m_lStartOffset * 1000 / 75;
   m_SeekTime = iTime;
-  m_callback.OnPlayBackSeek(m_SeekTime);
+  m_callback.OnPlayBackSeek((int)m_SeekTime);
   CLog::Log(LOGDEBUG, "PAPlayer::Seeking to time %f", 0.001f * m_SeekTime);
 }
